Fix toggle visible
