+++ crates/gpui/examples/tray.rs
@@ -1,0 +1,152 @@
+use gpui::{
+    App, Application, Context, Global, Image, MenuItem, SharedString, Tray, Window, WindowOptions,
+    actions, div, prelude::*,
+};
+
+struct Example;
+
+impl Render for Example {
+    fn render(&mut self, _window: &mut Window, _cx: &mut Context<Self>) -> impl IntoElement {
+        div()
+            .bg(gpui::white())
+            .flex()
+            .flex_col()
+            .gap_4()
+            .size_full()
+            .justify_center()
+            .items_center()
+            .child("Example for set Tray Icon")
+            .child(
+                div()
+                    .id("toggle-visible")
+                    .py_0p5()
+                    .px_3()
+                    .bg(gpui::black())
+                    .active(|this| this.bg(gpui::black().opacity(0.8)))
+                    .text_color(gpui::white())
+                    .child("Toggle Tray Visibility")
+                    .on_click(|_, window, cx| {
+                        window.dispatch_action(Box::new(ToggleVisible), cx);
+                    }),
+            )
+    }
+}
+
+fn main() {
+    Application::new().run(|cx: &mut App| {
+        cx.set_global(AppState::new());
+
+        // Bring the menu bar to the foreground (so you can see the menu bar)
+        cx.activate(true);
+        // Register the `quit` function so it can be referenced by the `MenuItem::action` in the menu bar
+        cx.on_action(quit);
+        cx.on_action(toggle_check);
+        cx.on_action(toggle_visible);
+
+        cx.open_window(WindowOptions::default(), |_, cx| cx.new(|_| Example))
+            .unwrap();
+
+        let app_state = cx.global::<AppState>();
+        cx.set_tray(app_state.tray.clone());
+    });
+}
+
+#[derive(PartialEq)]
+enum ViewMode {
+    List,
+    Grid,
+}
+
+impl ViewMode {
+    fn as_str(&self) -> &'static str {
+        match self {
+            ViewMode::List => "List",
+            ViewMode::Grid => "Grid",
+        }
+    }
+
+    fn toggle(&mut self) {
+        *self = match self {
+            ViewMode::List => ViewMode::Grid,
+            ViewMode::Grid => ViewMode::List,
+        }
+    }
+}
+
+impl Into<SharedString> for ViewMode {
+    fn into(self) -> SharedString {
+        match self {
+            ViewMode::List => "List",
+            ViewMode::Grid => "Grid",
+        }
+        .into()
+    }
+}
+
+struct AppState {
+    view_mode: ViewMode,
+    tray: Tray,
+}
+
+impl AppState {
+    fn new() -> Self {
+        Self {
+            view_mode: ViewMode::List,
+            tray: Tray::new()
+                .icon(Image::from_bytes(
+                    gpui::ImageFormat::Png,
+                    include_bytes!("image/app-icon.png").to_vec(),
+                ))
+                .title("Tray App")
+                .tooltip("This is a tray icon")
+                .menu(Self::build_menus),
+        }
+    }
+
+    fn build_menus(cx: &mut App) -> Vec<MenuItem> {
+        let app_state = cx.global::<AppState>();
+
+        vec![
+            MenuItem::action(ViewMode::List, ToggleCheck)
+                .checked(app_state.view_mode == ViewMode::List),
+            MenuItem::action(ViewMode::Grid, ToggleCheck)
+                .checked(app_state.view_mode == ViewMode::Grid),
+            MenuItem::separator(),
+            MenuItem::action("Hide Tray Icon", ToggleVisible),
+            MenuItem::separator(),
+            MenuItem::action("Quit", Quit),
+        ]
+    }
+}
+
+impl Global for AppState {}
+
+// Associate actions using the `actions!` macro (or `Action` derive macro)
+actions!(example, [Quit, ToggleCheck, ToggleVisible]);
+
+// Define the quit function that is registered with the App
+fn quit(_: &Quit, cx: &mut App) {
+    println!("Gracefully quitting the application . . .");
+    cx.quit();
+}
+
+fn toggle_check(_: &ToggleCheck, cx: &mut App) {
+    println!("Toggling view mode . . .");
+
+    {
+        let app_state = cx.global_mut::<AppState>();
+        app_state.view_mode.toggle();
+        app_state.tray.title = Some(format!("Mode: {}", app_state.view_mode.as_str()).into());
+    }
+
+    let app_state = cx.global::<AppState>();
+    cx.set_tray(app_state.tray.clone());
+}
+
+fn toggle_visible(_: &ToggleVisible, cx: &mut App) {
+    let app_state = cx.global_mut::<AppState>();
+    app_state.tray.visible = !app_state.tray.visible;
+
+    let app_state = cx.global::<AppState>();
+    cx.set_tray(app_state.tray.clone());
+}
