+++ crates/gpui/src/platform/mac.rs
@@ -5,7 +5,7 @@ mod display;
 mod display_link;
 mod events;
 mod keyboard;
-mod system_tray;
+mod tray;
 
 #[cfg(feature = "screen-capture")]
 mod screen_capture;
@@ -51,6 +51,7 @@ pub(crate) use display::*;
 pub(crate) use display_link::*;
 pub(crate) use keyboard::*;
 pub(crate) use platform::*;
+pub(crate) use tray::*;
 pub(crate) use window::*;
 
 #[cfg(feature = "font-kit")]
