+++ crates/gpui/src/tray.rs
@@ -4,15 +4,20 @@ use crate::{App, Image, MenuItem, SharedString};
 
 /// System tray icon.
 #[derive(Clone)]
-pub struct SystemTray {
-    pub(crate) tooltip: Option<SharedString>,
-    pub(crate) title: Option<SharedString>,
-    pub(crate) icon: Option<Rc<Image>>,
-    pub(crate) menu_builder: Option<Rc<dyn Fn(&mut App) -> Vec<MenuItem>>>,
-    pub(crate) visible: bool,
+pub struct Tray {
+    /// Tooltip text.
+    pub tooltip: Option<SharedString>,
+    /// Tray title after the Icon.
+    pub title: Option<SharedString>,
+    /// Tray icon image.
+    pub icon: Option<Rc<Image>>,
+    /// Function to build the context menu.
+    pub menu_builder: Option<Rc<dyn Fn(&mut App) -> Vec<MenuItem>>>,
+    /// Visibility of the tray icon.
+    pub visible: bool,
 }
 
-impl SystemTray {
+impl Tray {
     /// Create a new tray icon with default properties.
     pub fn new() -> Self {
         Self {
@@ -55,20 +60,5 @@ impl SystemTray {
     pub fn visible(mut self, visible: bool) -> Self {
         self.visible = visible;
         self
-    }
-
-    /// Set the tooltip text.
-    pub fn set_tooltip(&mut self, tooltip: impl Into<SharedString>) {
-        self.tooltip = Some(tooltip.into());
-    }
-
-    /// Set the title text.
-    pub fn set_title(&mut self, title: impl Into<SharedString>) {
-        self.title = Some(title.into());
-    }
-
-    /// Set the icon image.
-    pub fn set_icon(&mut self, icon: Image) {
-        self.icon = Some(Rc::new(icon));
     }
 }
